[1.0.0] - add the browser version and keep the old node one
